+++ src/mpmToolbox/gui/syncPlayer/SyncPlayer.java
@@ -1,8 +1,9 @@
 package mpmToolbox.gui.syncPlayer;
 
+import com.alee.api.annotations.NotNull;
 import com.alee.laf.button.WebButton;
-import com.alee.laf.checkbox.WebCheckBox;
 import com.alee.laf.combobox.WebComboBox;
+import com.alee.laf.label.WebLabel;
 import com.alee.laf.panel.WebPanel;
 import com.alee.laf.slider.WebSlider;
 import com.alee.laf.spinner.WebSpinner;
@@ -10,12 +11,15 @@ import meico.audio.Audio;
 import meico.audio.AudioPlayer;
 import meico.midi.Midi;
 import meico.midi.MidiPlayer;
+import meico.mpm.elements.Performance;
+import meico.supplementary.KeyValue;
 import mpmToolbox.gui.ProjectPane;
 import mpmToolbox.gui.Settings;
 import mpmToolbox.supplementary.Tools;
 
 import javax.sound.midi.InvalidMidiDataException;
 import javax.sound.midi.MidiUnavailableException;
+import javax.sound.sampled.Clip;
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseEvent;
@@ -27,23 +31,23 @@ import java.awt.event.MouseListener;
  * @author Axel Berndt
  */
 public class SyncPlayer extends WebPanel {
-    private final ProjectPane parent;                               // a link to the parent project pane to access its data, midi player etc.
+    private final ProjectPane parent;                                   // a link to the parent project pane to access its data, midi player etc.
 
-    private WebButton playButton = new WebButton("\u25B6");         //  ◼ "\u25FC", ▶ "\u25B6"
+    private final WebButton playButton = new WebButton("\u25B6");       //  ◼ "\u25FC", ▶ "\u25B6"
     private static final int sliderMax = 1000000;
-    private WebSlider playbackSlider = new WebSlider(WebSlider.HORIZONTAL, 0, sliderMax, 0);  // the slider that indicates playback position
+    private final WebSlider playbackSlider = new WebSlider(WebSlider.HORIZONTAL, 0, sliderMax, 0);  // the slider that indicates playback position
+    private boolean sliderBlocked = false;
 
-    private WebCheckBox playMidi = new WebCheckBox("Play Performance Rendering");   // TODO: to be deleted
-    private WebCheckBox playAudio = new WebCheckBox("Play Audio");                  // TODO: to be deleted
+    private final AudioPlayer audioPlayer = new AudioPlayer();
+    private final MidiPlayer midiPlayer = new MidiPlayer();
 
-    private WebComboBox performanceChooser = new WebComboBox();     // TODO: should include "None" and "Raw MIDI/No Performance"
-    private WebComboBox audioChooser = new WebComboBox();           // TODO: should include "None"
+    private final WebComboBox performanceChooser = new WebComboBox();
+    private final PerformanceChooserItem rawPerformance = new PerformanceChooserItem(Performance.createPerformance("Play raw notes, no performance"));
 
-    private WebSpinner skipMillisecondsInAudioPlayback = new WebSpinner();
-    private AudioPlayer audioPlayer = new AudioPlayer();
-    private MidiPlayer midiPlayer = new MidiPlayer();
+    private final WebComboBox audioChooser = new WebComboBox();
+    private final WebSpinner skipMillisecondsInAudioPlayback = new WebSpinner(new SpinnerNumberModel(0L, 0L, 9999999999L, 1L));
 
-    private PlaybackRunnable runnable = new PlaybackRunnable();
+    private PlaybackRunnable runnable = null;
 
     /**
      * constructor
@@ -74,17 +78,72 @@ public class SyncPlayer extends WebPanel {
      * this creates the Sync Player GUI
      */
     private void makeGui() {
-        this.makeCheckboxes();
+        this.updatePerformanceList();
+        this.performanceChooser.setPadding(Settings.paddingInDialogs / 4);
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.performanceChooser, 0, 0, 4, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
+        this.updateAudioList();
+        this.audioChooser.setPadding(Settings.paddingInDialogs / 4);
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.audioChooser, 0, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
+        WebLabel skipLabel = new WebLabel("skip");
+        skipLabel.setToolTip("Skip initial silence in the audio recording.");
+        skipLabel.setHorizontalAlignment(WebLabel.CENTER);
+        skipLabel.setPadding(Settings.paddingInDialogs / 4);
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), skipLabel, 1, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.skipMillisecondsInAudioPlayback, 2, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
+        WebLabel millisecondsLabel = new WebLabel("ms");
+        millisecondsLabel.setToolTip("Milliseconds");
+        millisecondsLabel.setHorizontalAlignment(WebLabel.LEFT);
+        millisecondsLabel.setPadding(Settings.paddingInDialogs / 4);
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), millisecondsLabel, 3, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
+        this.playButton.setPadding((int) (Settings.paddingInDialogs * 1.5));
+        this.playButton.addActionListener(actionEvent -> this.triggerPlayback());
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playButton, 4, 0, 1, 2, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+
         this.makeSlider();
-        this.makePlayButton();
     }
 
     /**
-     * add the cehckboxes for choosing the playback source to the GUI
+     * This fills the performance chooser list
+     * It method should be called when the the available performances have changed.
      */
-    private void makeCheckboxes() {
-        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playMidi, 0, 0, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playAudio, 0, 1, 1, 1, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+    public void updatePerformanceList() {
+        PerformanceChooserItem selectedItem = (PerformanceChooserItem) this.performanceChooser.getSelectedItem();   // store the previously selected item
+
+        this.performanceChooser.removeAllItems();
+        this.performanceChooser.addItem(new PerformanceChooserItem("No performance rendering"));
+        this.performanceChooser.addItem(this.rawPerformance);
+        if (this.parent.getMpm() != null) {
+            for (Performance performance : this.parent.getMpm().getAllPerformances()) {
+                PerformanceChooserItem item = new PerformanceChooserItem(performance);
+                this.performanceChooser.addItem(item);
+
+                if ((selectedItem != null) && item.toString().equals(selectedItem.toString()))  // if this is the item that was previously selected
+                    this.performanceChooser.setSelectedItem(item);                              // keep this selection
+            }
+        }
+    }
+
+    /**
+     * This fills the audio chooser list.
+     * It should be called when the list of audio recordings changes.
+     */
+    public void updateAudioList() {
+        AudioChooserItem selectedItem = (AudioChooserItem) this.audioChooser.getSelectedItem(); // store the previously selected item for reference
+
+        this.audioChooser.removeAllItems();
+        this.audioChooser.addItem(new AudioChooserItem("No audio recording"));
+        for (Audio audio : this.parent.getAudio()) {
+            AudioChooserItem item = new AudioChooserItem(audio);
+            this.audioChooser.addItem(item);
+
+            if ((selectedItem != null) && item.toString().equals(selectedItem.toString()))      // if this is the item that was previously selected
+                this.audioChooser.setSelectedItem(item);                                        // keep this selection
+        }
     }
 
     /**
@@ -94,17 +153,23 @@ public class SyncPlayer extends WebPanel {
         this.playbackSlider.setMajorTickSpacing(sliderMax / 4);
         this.playbackSlider.setMinorTickSpacing(sliderMax / 16);
         this.playbackSlider.setPaintTicks(true);
+        Tools.makeSliderSetToClickPosition(this.playbackSlider);
 
+        // define interaction
         this.playbackSlider.addMouseListener(new MouseListener() {
             @Override
             public void mouseClicked(MouseEvent mouseEvent) {
-                // TODO: set playback position in midi and audio
             }
             @Override
             public void mousePressed(MouseEvent mouseEvent) {
+                sliderBlocked = true;   // don't allow the runnable to change the slider position while the user interacts with it
             }
             @Override
             public void mouseReleased(MouseEvent mouseEvent) {
+                if (runnable != null) {                             // if music is already playing
+                    runnable.jumpTo(((double) playbackSlider.getValue()) / sliderMax);
+                }
+                sliderBlocked = false;  // now the runnable is allowed again to change the slider position
             }
             @Override
             public void mouseEntered(MouseEvent mouseEvent) {
@@ -114,70 +179,23 @@ public class SyncPlayer extends WebPanel {
             }
         });
 
-        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playbackSlider, 2, 0, 1, 2, 100.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
+        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playbackSlider, 5, 0, 1, 2, 100.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
     }
 
     /**
-     * helper method to define the play button
+     * start/stop playback
      */
-    private void makePlayButton() {
-//        this.playButton = new WebButton(this.getMidiPlayer().isPlaying() ? "\u25FC" : "\u25B6");    //  ◼ "\u25FC", ▶ "\u25B6"
-        this.playButton.setPadding(Settings.paddingInDialogs);
-//        this.playButton.setFontSize(Settings.getDefaultFontSize());
-        this.playButton.addActionListener(actionEvent -> this.triggerPlayback());   // set the button's action
-        Tools.addComponentToGridBagLayout(this, (GridBagLayout) this.getLayout(), this.playButton, 1, 0, 1, 2, 1.0, 1.0, 0, 0, GridBagConstraints.BOTH, GridBagConstraints.LINE_START);
-    }
-
-    /**
-     * start the playback or stop it if it is already running
-     */
-    private void triggerPlayback() {
-        // if music is already playing, we only want to stop playback
-        if (this.getMidiPlayer().isPlaying() || this.getAudioPlayer().isPlaying()) {
-            this.playButton.setText("\u25B6");              // set the playButton's symbol to ▶
-            this.runnable.terminate();                      // terminate the current runnable/thread, this will also stop the players
+    public synchronized void triggerPlayback() {
+        if ((this.runnable != null) && (this.runnable.isPlaying())) {   // if music is already playing, we only want to stop it
+            this.playButton.setText("\u25B6");                          // set the playButton's symbol to ▶
+            this.runnable.stop();                                       // terminate the current runnable/thread, this will also stop the players
+            this.runnable = null;
             return;
         }
 
         // we want to start a new playback
-        double relativeSliderValue = ((double) this.playbackSlider.getValue()) / sliderMax;
-        long playbackPosition = 0;
-
-        Midi midi = null;
-        if (this.playMidi.isSelected()) {
-            midi = (this.parent.getMpm() == null) ? this.parent.getMsm().exportMidi(100) : this.parent.getMsm().exportExpressiveMidi(this.parent.getMpm().getPerformance(0), true); // render the MIDI performance, TODO: choose the right performance
-            playbackPosition = (long) ((double) midi.getMicrosecondLength() * relativeSliderValue);
-        }
-
-        if (this.playAudio.isSelected() && !this.parent.getAudio().isEmpty()) {
-            Audio audio = this.parent.getAudio().get(0);                                                        // set the audio data to be played back, // TODO: choose the right audio
-            if (this.getAudioPlayer().setAudioData(audio)) {                                                    // TODO: make sure we have non-null audio data
-                if ((midi == null) || (midi.getMicrosecondLength() < this.getAudioPlayer().getMicrosecondLength())) {// the slider's length must cover the longer of both, MIDI and Audio, so if Audio is longer, update the playback position
-                    playbackPosition = (long) ((double) this.getAudioPlayer().getMicrosecondLength() * relativeSliderValue);
-                }
-                this.getAudioPlayer().setMicrosecondPosition(playbackPosition);
-                this.getAudioPlayer().play();
-            }
-        }
-
-        if (midi != null) {
-            double relativeMidiPlaybackPosition = (double) playbackPosition / midi.getMicrosecondLength();
-            if (relativeMidiPlaybackPosition < 1.0) {
-                try {
-                    this.getMidiPlayer().play(midi, relativeMidiPlaybackPosition);   // start MIDI playback at the slider position
-                } catch (InvalidMidiDataException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-
-        if (this.getMidiPlayer().isPlaying() || this.getAudioPlayer().isPlaying()) {
-            this.playButton.setText("\u25FC");                  // set the playButton's symbol
-
-            this.runnable.terminate();                          // terminate the current runnable/thread (if there was none running, this is superficial but it does not hurt either)
-            this.runnable = new PlaybackRunnable();             // create a new runnable
-            (new Thread(this.runnable)).start();                // initialize and start a new thread with that new runnable
-        }
+        this.runnable = new PlaybackRunnable();
+        this.runnable.start(((double) this.playbackSlider.getValue()) / sliderMax); // start the new runnable
     }
 
     /**
@@ -187,26 +205,128 @@ public class SyncPlayer extends WebPanel {
      * @author Axel Berndt
      */
     private class PlaybackRunnable implements Runnable {
+        private volatile Thread thread = null;
         private volatile boolean terminate = false;
+
+        private double relativeMidiPlaybackPosition = 0.0;
+        private long audioPlaybackPosition = 0;
+        private long microsecAudioOffset = 0;
+
+        private final Midi midi;
+        private final Clip audio;
+        private final boolean midiIsLonger;
+
+        /**
+         * constructor
+         */
+        protected PlaybackRunnable() {
+            if ((performanceChooser.getSelectedItem() == null) || (((PerformanceChooserItem) performanceChooser.getSelectedItem()).getValue() == null))
+                this.midi = null;
+            else
+                this.midi = parent.getMsm().exportExpressiveMidi(((PerformanceChooserItem) performanceChooser.getSelectedItem()).getValue(), true);
+
+            getAudioPlayer().stop();
+            if ((audioChooser.getSelectedItem() != null) && getAudioPlayer().setAudioData(((AudioChooserItem) audioChooser.getSelectedItem()).getValue())) {
+                this.audio = getAudioPlayer().getAudioClip();
+                this.microsecAudioOffset = (long) ((double) skipMillisecondsInAudioPlayback.getValue()) * 1000;
+            }
+            else
+                this.audio = null;
+
+            this.midiIsLonger = (this.audio == null) || ((this.midi != null) && (this.midi.getMicrosecondLength() > (this.audio.getMicrosecondLength() - this.microsecAudioOffset)));
+        }
+
+        /**
+         * initialize and start a new thread with this runnable
+         * @param relativeSliderPosition
+         */
+        protected void start(double relativeSliderPosition) {
+            if (((this.audio == null) && (this.midi == null)) || (this.thread != null))
+                return;
+
+            playButton.setText("\u25FC");                      // set the playButton's symbol to ◼
+            this.setPlaybackPositions(relativeSliderPosition);
+            this.thread = new Thread(this);
+            this.thread.start();
+        }
+
+        /**
+         * start those players that have data
+         */
+        private void startPlayers() {
+            if ((this.midi != null) && (this.relativeMidiPlaybackPosition < 1.0)) {
+                try {
+                    getMidiPlayer().play(this.midi, this.relativeMidiPlaybackPosition);   // start MIDI playback at the slider position
+                } catch (InvalidMidiDataException e) {
+                    e.printStackTrace();
+                }
+            }
+
+            if ((this.audio != null) && (this.audioPlaybackPosition < getAudioPlayer().getMicrosecondLength())) {
+                getAudioPlayer().setMicrosecondPosition(this.audioPlaybackPosition);
+                getAudioPlayer().play();
+            }
+        }
 
         /**
          * invoke this method to terminate the runnable and, thus, the thread
          */
-        protected synchronized void terminate() {
+        protected synchronized void stop() {
+            getMidiPlayer().stop();
+            getAudioPlayer().stop();
             this.terminate = true;
         }
 
         /**
-         * If the audio playback reached the end, it does not stop the player automatically.
-         * So we have to do it explicitly. That is what this method does.
-         * @return true when stopped
+         * signal the runnable to jump to another playback position
+         * @param relativeSliderPosition
          */
-        private boolean audioAutoStop() {
-            if (getAudioPlayer().getRelativePosition() >= 1.0) {
-                getAudioPlayer().stop();
-                return true;
+        protected synchronized void jumpTo(double relativeSliderPosition) {
+            if (!this.isPlaying())
+                return;
+
+            this.setPlaybackPositions(relativeSliderPosition);
+
+            if (this.audio != null) {
+                getAudioPlayer().pause();
+                getAudioPlayer().setAudioData(this.audio);
+                getAudioPlayer().setMicrosecondPosition(this.audioPlaybackPosition);
+                getAudioPlayer().play();
             }
-            return false;
+
+            if (this.midi != null) {
+                try {
+                    getMidiPlayer().play(this.midi, this.relativeMidiPlaybackPosition);
+                } catch (InvalidMidiDataException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+        /**
+         * set the playback positions before starting playback
+         * @param relativeSliderPosition
+         */
+        private void setPlaybackPositions(double relativeSliderPosition) {
+            if (this.midiIsLonger) {
+                this.relativeMidiPlaybackPosition = relativeSliderPosition;
+                if (this.audio != null)
+                    this.audioPlaybackPosition = (long) (relativeSliderPosition * this.midi.getMicrosecondLength()) + this.microsecAudioOffset;
+            }
+            else {    // audio is longer
+                long effectiveAudioLength = getAudioPlayer().getMicrosecondLength() - this.microsecAudioOffset;
+                this.audioPlaybackPosition = (long) (relativeSliderPosition * effectiveAudioLength) + this.microsecAudioOffset;
+                if (this.midi != null)
+                    this.relativeMidiPlaybackPosition = ((double) this.audioPlaybackPosition - this.microsecAudioOffset) / this.midi.getMicrosecondLength();
+            }
+        }
+
+        /**
+         * indicates whether the playback ist still running
+         * @return
+         */
+        protected boolean isPlaying() {
+            return getAudioPlayer().isPlaying() || getMidiPlayer().isPlaying();
         }
 
         /**
@@ -214,22 +334,25 @@ public class SyncPlayer extends WebPanel {
          */
         @Override
         public void run() {
-            while (!terminate) {  // while the music plays
+            this.startPlayers();
+
+            while (!this.terminate) {  // while the music plays
 //                this.playButton.setText("\u25FC");
 
                 // update playbackSlider
                 double relativePlaybackPosition;
-                if (getAudioPlayer().isPlaying() && getMidiPlayer().isPlaying()) {
-                    relativePlaybackPosition = (getAudioPlayer().getMicrosecondLength() > getMidiPlayer().getMicrosecondLength()) ? getAudioPlayer().getRelativePosition() : getMidiPlayer().getRelativePosition();
-                    this.audioAutoStop();
-                } else if (getAudioPlayer().isPlaying()) {
-                    if (this.audioAutoStop())
-                        break;
-                    relativePlaybackPosition = getAudioPlayer().getRelativePosition();
-                } else if (getMidiPlayer().isPlaying())
+                if (this.midiIsLonger)
                     relativePlaybackPosition = getMidiPlayer().getRelativePosition();
-                else break;
-                playbackSlider.setValue((int) (relativePlaybackPosition * sliderMax));
+                else
+                    relativePlaybackPosition = (double) (getAudioPlayer().getMicrosecondPosition() - this.microsecAudioOffset) / (double) (getAudioPlayer().getMicrosecondLength() - this.microsecAudioOffset);
+
+                if ((playbackSlider.getValue() == sliderMax) || (!getAudioPlayer().isPlaying() && !getMidiPlayer().isPlaying())) {
+                    runnable = null;
+                    break;
+                }
+
+                if (!sliderBlocked)
+                    playbackSlider.setValue((int) (relativePlaybackPosition * sliderMax));
 
                 try {
                     Thread.sleep(200L);
@@ -240,14 +363,79 @@ public class SyncPlayer extends WebPanel {
 
             // playback stops, this thread must terminate, but before that, do some housekeeping
             SwingUtilities.invokeLater(() -> {      // GUI operations must be done on the Event Dispatch Thread
-                getMidiPlayer().stop();
-                getAudioPlayer().stop();
-
-                if (!terminate)                     // if we reached the end of the music, i.e. playback was not terminated by interaction
+                if (!terminate && !sliderBlocked)   // if we reached the end of the music, i.e. playback was not terminated by interaction
                     playbackSlider.setValue(0);     // set slider to start position ... in any other case just keep the slider position
 
                 playButton.setText("\u25B6");       //  ▶ "\u25B6"
             });
+        }
+    }
+
+    /**
+     * This class represents an item in the performance chooser combobox of the SyncPlayer.
+     * @author Axel Berndt
+     */
+    private class PerformanceChooserItem extends KeyValue<String, Performance> {
+        /**
+         * This constructor creates a performance chooser item (String, Performance) pair out of a non-null performance.
+         * @param performance
+         */
+        private PerformanceChooserItem(@NotNull Performance performance) {
+            super(performance.getName(), performance);
+        }
+
+        /**
+         * This constructor creates a performance chooser item with the specified name key but null performance.
+         * Basically, this is used to communicate to the SyncPlayer not to play a performance rendering.
+         * The string is typically something like "No performance rendering".
+         * @param string
+         */
+        private PerformanceChooserItem(String string) {
+            super(string, null);
+        }
+
+        /**
+         * All combobox items require this method. The overwrite here makes sure that the string being returned
+         * is the performance's name instead of some Java Object ID.
+         * @return
+         */
+        @Override
+        public String toString() {
+            return this.getKey();
+        }
+    }
+
+    /**
+     * This class represents an item in the audio chooser combobox of the SyncPlayer.
+     * @author Axel Berndt
+     */
+    private class AudioChooserItem extends KeyValue<String, Audio> {
+        /**
+         * This constructor creates a audio chooser item (String, Audio) pair out of a non-null audio object.
+         * @param audio
+         */
+        private AudioChooserItem(@NotNull Audio audio) {
+            super(audio.getFile().getName(), audio);
+        }
+
+        /**
+         * This constructor creates a audio chooser item with the specified name key but null audio object.
+         * Basically, this is used to communicate to the SyncPlayer not to play audio.
+         * The string is typically something like "No audio recording".
+         * @param string
+         */
+        private AudioChooserItem(String string) {
+            super(string, null);
+        }
+
+        /**
+         * All combobox items require this method. The overwrite here makes sure that the string being returned
+         * is the audio file's name instead of some Java Object ID.
+         * @return
+         */
+        @Override
+        public String toString() {
+            return this.getKey();
         }
     }
 }